jangan lupa push

+++ database/migrations/2022_01_13_022655_create_profesis_table.php
@@ -20,7 +20,7 @@ class CreateProfesisTable extends Migration
             $table->bigInteger('price_low');
             $table->bigInteger('price_medium');
             $table->bigInteger('price_high');
-            $table->integer('margin');
+            $table->bigInteger('margin');
             $table->timestamps();
         });
     }
